Adjust redirect behavior after load
