+++ recce/apis/check_api.py
@@ -2,15 +2,13 @@ from datetime import datetime
 from typing import Optional
 from uuid import UUID
 
-from fastapi import APIRouter, HTTPException, UploadFile
-from fastapi.responses import PlainTextResponse
-from pydantic import BaseModel, ValidationError
+from fastapi import APIRouter, HTTPException
+from pydantic import BaseModel
 
 from recce.apis.check_func import get_node_by_id, validate_schema_diff_check
 from recce.apis.run_func import submit_run
 from recce.exceptions import RecceException
 from recce.models import RunType, RunDAO, Check, CheckDAO, Run
-from recce.models.util import pydantic_model_json_dump
 
 check_router = APIRouter(tags=['check'])
 
@@ -232,32 +230,5 @@ class ReorderChecksIn(BaseModel):
 async def reorder_handler(order: ReorderChecksIn):
     try:
         CheckDAO().reorder(order.source, order.destination)
-    except RecceException as e:
-        raise HTTPException(status_code=400, detail=e.message)
-
-
-@check_router.post("/checks/export", response_class=PlainTextResponse, status_code=200)
-async def export_handler():
-    from ..models.state import recce_state
-
-    try:
-        return pydantic_model_json_dump(recce_state)
-    except RecceException as e:
-        raise HTTPException(status_code=400, detail=e.message)
-
-
-@check_router.post("/checks/load", status_code=200)
-async def load_handler(file: UploadFile):
-    from ..models.state import RecceState, recce_state
-
-    try:
-        content = await file.read()
-        load_state = RecceState().model_validate_json(content)
-        recce_state.checks = load_state.checks
-        recce_state.runs = load_state.runs
-
-        return {"runs": len(recce_state.runs), "checks": len(recce_state.checks)}
-    except ValidationError as e:
-        raise HTTPException(status_code=400, detail=str(e))
     except RecceException as e:
         raise HTTPException(status_code=400, detail=e.message)
